+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Fixed
 
 - Fix error reporting bug in `get_signal` and `signal_orange_to_ewoks_name`.
+- Hide Ewoks task inputs/outputs from Orange propagation when hidden from Orange.
 
 ## [4.0.0] - 2026-03-11